% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnist.R
\name{load_mnist}
\alias{load_mnist}
\title{Load the MNIST images}
\usage{
load_mnist(train = FALSE, download_if_missing = FALSE)
}
\arguments{
\item{train}{Logical, indicating if the training data set should
be loaded.  If \code{FALSE} then the "t10k" set is loaded.  By
default, the smaller file is loaded.}

\item{download_if_missing}{Logical, indicating if the images
should be downloaded if not present.  See
\code{\link{download_mnist}} for details about the downloading
and where files will be stored.}
}
\description{
Load the MNIST images.  There are two sets "train" and "t10k"; the
\code{train} argument switches between these.  The actual images
are not shipped with the package as they are ~11MB.  Instead these
can be downloaded with \code{download_mnist} or by passing
\code{download_if_missing = TRUE} to this function.
}
\details{
The first time within a session that the images are loaded, the
loading process may be slow as the files are read from disk (this
is particularly true for the "train = TRUE" data set, which is the
larger of the two).  Subsequent loads will be much faster though,
as the processed images are cached within a session.
}
\examples{
\donttest{
mnist <- load_mnist(FALSE)
mnist

x <- mnist[[2]]
x
print(x, TRUE)
plot(x)
}
}

