% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnist.R
\name{download_mnist}
\alias{download_mnist}
\title{Download the MNIST images}
\usage{
download_mnist(verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{verbose}{Print a message even if the images are already found}

\item{quiet}{Passed through to \code{download.file} to suppress
the download progress bar.}
}
\description{
Download the MNIST images.  Because these about 11MB they do not
ship with the package, but are downloaded as needed.
}
\details{
The downloaded images will be stored on your system by default at
the directory given by \code{rappdirs::user_cache_dir("rmnist")}.
Alternatively, you can specify your own location for the images by
setting the option \code{rmnist.cache_dir} (e.g.,
\code{rmnist.cache_dir = tempfile()}).  To see what value will be
used, you can run the (unexported) function
\code{rmnist:::cache_dir()}.
}

